Fix all ruff findings across the codebase
- Apply ruff autofixes: unused imports, import sorting, if-stmt-min-max,
  if-with-same-arms
- Make shebang scripts executable (EXE001)
- Use enumerate() for loop counter (SIM113)
- Return conditions directly, drop redundant int() casts (SIM103, RUF046)
- Iterate dict keys directly (PERF102)
- Use timezone-aware date.today() replacements (DTZ011)
- Add ruff.toml ignoring marimo cell-format false positives (B018, PLR1711)
- Targeted noqa for intentional naive-UTC date parsing and catch-all
  availability fetch

+++ ged-analysis.py
@@ -53,7 +53,7 @@ def geodesic_circle(lon, lat, radius_km,
     radius_m = radius_km * 1000.0
     tol = min(float(tolerance_m), radius_m)
     half_angle = 2 * math.asin(math.sqrt(tol / (2 * radius_m)))
-    n = max(int(math.ceil(math.pi / half_angle)), 8)
+    n = max(math.ceil(math.pi / half_angle), 8)
     azimuths = [360.0 * i / n for i in range(n)]
     lons, lats, _ = GEOD.fwd([float(lon)] * n, [float(lat)] * n,
                              azimuths, [radius_m] * n)
